requirements.txt
shotgun_api3>=3.1.4
pyexiftool>=0.5.0
ffmpeg-python>=0.2.0
PyQt5>=5.15.4
PyYAML>=6.0
pillow>=9.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
# For converting README.md to HTML
markdown
